+++ apps/backend/src/modules/ai-assistant/libs/validation-schemas/validation-schemas.ts
@@ -1,4 +1,5 @@
 export {
+	acceptMultipleTasksValidationSchema,
 	addMessageToThreadValidationSchema,
 	changeTaskSuggestionRequestValidationSchema,
 	taskActionRequestSchemaValidationSchema,