+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/helpers/get-dot-offset-y/get-dot-offset-y.helper.ts
@@ -1,21 +1,19 @@
-import { DEFAULT_GRAPHICS_OFFSET } from "../../constants/constants.js";
-import { DotOffsetY } from "../../enums/enums.js";
+import {
+	DEFAULT_GRAPHICS_OFFSET,
+	RIGHTMOST_DOTS_OFFSET_Y,
+} from "../../constants/constants.js";
 import { type GetGraphicsCoordinateOffset } from "../../types/types.js";
 
 const getDotOffsetY: GetGraphicsCoordinateOffset = (
 	graphicsPosition,
 	centers,
 ) => {
-	const { x, y } = graphicsPosition;
+	const { x } = graphicsPosition;
 
-	const { secondPartCenterX, secondPartCenterY } = centers;
+	const { secondPartCenterX } = centers;
 
 	if (x > secondPartCenterX) {
-		return DotOffsetY.RIGHTMOST;
-	}
-
-	if (y > secondPartCenterY) {
-		return DotOffsetY.BOTTOM_CENTER;
+		return RIGHTMOST_DOTS_OFFSET_Y;
 	}
 
 	return DEFAULT_GRAPHICS_OFFSET;