+++ apps/frontend/src/libs/enums/enums.ts
@@ -9,4 +9,5 @@ export {
 	ErrorMessage,
 	NumericalValue,
 	ServerErrorType,
+	ZERO_INDEX,
 } from "shared";