+++ apps/frontend/src/modules/categories/libs/types/types.ts
@@ -2,4 +2,5 @@ export {
 	type CategoriesGetAllResponseDto,
 	type CategoriesGetRequestQueryDto,
 	type CategoryDto,
+	type SelectedCategory,
 } from "shared";