+++ apps/frontend/src/modules/onboarding/onboarding-api.ts
@@ -1,7 +1,6 @@
 import { APIPath, ContentType } from "~/libs/enums/enums.js";
 import { BaseHTTPApi } from "~/libs/modules/api/api.js";
-import { type HTTP } from "~/libs/modules/http/http.js";
-import { type Storage } from "~/libs/modules/storage/storage.js";
+import { type APIConfiguration } from "~/libs/types/types.js";
 
 import { OnboardingApiPath } from "./libs/enums/enums.js";
 import {
@@ -10,14 +9,8 @@ import {
 	type OnboardingUserAnswerDto,
 } from "./libs/types/types.js";
 
-type Constructor = {
-	baseUrl: string;
-	http: HTTP;
-	storage: Storage;
-};
-
 class OnboardingApi extends BaseHTTPApi {
-	public constructor({ baseUrl, http, storage }: Constructor) {
+	public constructor({ baseUrl, http, storage }: APIConfiguration) {
 		super({ baseUrl, http, path: APIPath.ONBOARDING, storage });
 	}