+++ apps/frontend/src/modules/quiz/libs/types/types.ts
@@ -7,5 +7,6 @@ export {
 	type QuizScoresGetAllResponseDto,
 	type QuizScoresResponseDto,
 	type QuizScoresUpdateRequestDto,
+	type QuizScoresUpdateResponseDto,
 	type QuizUserAnswerDto,
 } from "shared";