+++ apps/frontend/src/modules/tasks/slices/actions.ts
@@ -1,9 +1,28 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
+import { NotificationMessage } from "~/libs/enums/enums.js";
 import { type AsyncThunkConfig } from "~/libs/types/types.js";
-import { type TaskDto, type TaskUpdatePayload } from "~/modules/tasks/tasks.js";
+import {
+	type TaskDto,
+	type TaskNoteDto,
+	type TaskNoteParametersDto,
+	type TaskNoteRequestDto,
+	type TaskUpdatePayload,
+} from "~/modules/tasks/tasks.js";
 
 import { name as sliceName } from "./tasks.slice.js";
+
+const addNote = createAsyncThunk<
+	TaskNoteDto,
+	TaskNoteRequestDto,
+	AsyncThunkConfig
+>(`${sliceName}/add-task-note`, async (payload, { extra }) => {
+	const { notification, tasksApi } = extra;
+
+	notification.success(NotificationMessage.TASK_NOTES_ADDED);
+
+	return await tasksApi.addNote(payload);
+});
 
 const getCurrentTasks = createAsyncThunk<
 	TaskDto[],
@@ -24,6 +43,17 @@ const getPastTasks = createAsyncThunk<TaskDto[], undefined, AsyncThunkConfig>(
 	},
 );
 
+const getTaskNotes = createAsyncThunk<
+	TaskNoteDto[],
+	TaskNoteParametersDto,
+	AsyncThunkConfig
+>(`${sliceName}/get-task-notes`, async (payload, { extra }) => {
+	const { tasksApi } = extra;
+	const { id } = payload;
+
+	return await tasksApi.getTaskNotes(id);
+});
+
 const update = createAsyncThunk<TaskDto, TaskUpdatePayload, AsyncThunkConfig>(
 	`${sliceName}/update`,
 	async (payload, { extra }) => {
@@ -43,4 +73,11 @@ const updateTaskDeadline = createAsyncThunk<TaskDto, number, AsyncThunkConfig>(
 	},
 );
 
-export { getCurrentTasks, getPastTasks, update, updateTaskDeadline };
+export {
+	addNote,
+	getCurrentTasks,
+	getPastTasks,
+	getTaskNotes,
+	update,
+	updateTaskDeadline,
+};