+++ apps/frontend/src/modules/tasks/slices/tasks.ts
@@ -1,6 +1,8 @@
 import {
+	addNote,
 	getCurrentTasks,
 	getPastTasks,
+	getTaskNotes,
 	update,
 	updateTaskDeadline,
 } from "./actions.js";
@@ -8,8 +10,10 @@ import { actions } from "./tasks.slice.js";
 
 const allActions = {
 	...actions,
+	addNote,
 	getCurrentTasks,
 	getPastTasks,
+	getTaskNotes,
 	update,
 	updateTaskDeadline,
 };