+++ packages/shared/src/index.ts
@@ -10,6 +10,7 @@ export {
 	ErrorMessage,
 	NumericalValue,
 	ServerErrorType,
+	SortOrder,
 } from "./libs/enums/enums.js";
 export {
 	AuthError,
@@ -23,7 +24,11 @@ export {
 	UserError,
 	ValidationError,
 } from "./libs/exceptions/exceptions.js";
-export { configureString, getFormattedDate } from "./libs/helpers/helpers.js";
+export {
+	configureString,
+	getFormattedDate,
+	getLocalDatestring,
+} from "./libs/helpers/helpers.js";
 export { type APIConfiguration } from "./libs/modules/api/api.js";
 export { type Config } from "./libs/modules/config/config.js";
 export {
@@ -44,8 +49,10 @@ export {
 	type ValueOf,
 } from "./libs/types/types.js";
 export {
+	acceptMultipleTasksValidationSchema,
 	addMessageToThreadValidationSchema,
 	AIAssistantApiPath,
+	type AIAssistantCreateMultipleTasksDto,
 	type AIAssistantRequestDto,
 	type AIAssistantResponseDto,
 	type AIAssistantSuggestTaskRequestDto,
@@ -72,7 +79,6 @@ export {
 	type SelectedCategory,
 } from "./modules/categories/categories.js";
 export {
-	type BalanceWheelMessage,
 	ChatMessageAuthor,
 	type ChatMessageDto,
 	ChatMessageType,
@@ -107,6 +113,7 @@ export {
 	type QuizScoresGetAllResponseDto,
 	type QuizScoresResponseDto,
 	type QuizScoresUpdateRequestDto,
+	type QuizScoresUpdateResponseDto,
 	type QuizUserAnswerDto,
 	quizUserAnswersValidationSchema,
 	updateScoresValidationSchema,
@@ -119,6 +126,7 @@ export {
 	type TaskNoteParametersDto,
 	type TaskNoteRequestDto,
 	taskNoteValidationSchema,
+	type TaskPayload,
 	TasksApiPath,
 	TaskStatus,
 	type TaskUpdateParametersDto,