+++ packages/shared/src/modules/users/libs/enums/user-validation-regex-rule.enum.ts
@@ -1,9 +1,10 @@
 import { UserValidationRule } from "./user-validation-rule.enum.js";
 
 const UserValidationRegexRule = {
-	EMAIL_DOMAIN_MIN_MAX_CHARS: /@.{4,35}$/,
-	EMAIL_DOMAIN_PART_VALID_CHARS: /(?<=@)[\dA-Za-z-]{4,35}(\.[\dA-Za-z-]{2,})+$/,
-	EMAIL_LOCAL_PART_VALID_CHARS: /^[\w!#$&'*+=\\-]{1,35}(?=@)/,
+	EMAIL_DOMAIN_PART_VALID_CHARS:
+		/^[^\s@]+@(?=.{4,35}$)[\dA-Za-z]+(?:-[\dA-Za-z]+)*\.[\dA-Za-z]{2,}$/,
+	EMAIL_LOCAL_PART_VALID_CHARS:
+		/^(?!\.)((?!\.{2})[\w!#$&'*+.=\\-]){1,35}(?<!\.)@/,
 	NAME_VALID_CHARS_MIN_MAX: new RegExp(
 		`^[a-zA-Z0-9_-]{${String(UserValidationRule.NAME_MIN_LENGTH)},${String(UserValidationRule.NAME_MAX_LENGTH)}}$`,
 	),